+++ src/modules/returns/lib/returns-upload-validator.js
@@ -24,7 +24,7 @@ const { flatMap, cond, negate, isEqual } = require('lodash')
 
 const returnsConnector = require('../../../lib/connectors/returns')
 const documents = require('../../../lib/connectors/crm/documents')
-const chunk = require('../../../lib/chunk.js')
+const { chunk } = require('../../../lib/object-helpers.js')
 
 const returnLines = require('@envage/water-abstraction-helpers').returns.lines
